Dashboard: Disable task retry on deferred tasks
This ensures that we don't get into potential multiple attempts at
failing tasks. This avoids the potential multi-filing of issues for a
given group if there are multiple tasks that are retrying concurrently.

In the future we probably want to use a mutex to prevent multiple of
these handlers from attempting to perform the triage.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1078838
Change-Id: I59dd607a5a70468c69294acdf2ec33d7c3d248c3