Reland "Do not update chrome_dev.conf on cros_browser backend"
This is a reland of e783d863d957423980d9f6ca559af18e596e7fa8

Original change was reverted due to the test failure of
CrOSBrowserEnvironmentTest.testMinidumpsFromTestRemoved. As I
debugged a bit, this is because SetUpEnvironment() becomes
a bit faster so that "/var/log/chrome/Crash Reports" does not
yet exist. This is not a problem on real environment, but
test invokes some commands a bit earlier than expected which
causes an error. So I added py_utils.WaitFor in the test case
to make sure the directory exists.

Original change's description:
> Do not update chrome_dev.conf on cros_browser backend
>
> cros_browser_finder requires writable rootfs to set environment
> variables on /etc/chrome_dev.conf. This is time consuming and
> causing troubles like crbug.com/1078418. The environment variables
> can be set through session manager, and actually the method is
> already invoked with empty extra environment variables.
>
> This CL stops updating /etc/chrome_dev.conf and channel the env
> data through the session manager.
>
> Bug: chromium:1078418
> Test: manually; run_benchmarks with a VM and see no reboots
> Change-Id: Ic564e7984456bd5d4e38a4996244e8e217f381aa
> Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2183205
> Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
> Commit-Queue: Jun Mukai <mukai@chromium.org>

Bug: chromium:1078418
Change-Id: Ice313ac57770e422ca451cc79f6b33b9b42f894b