[Telemetry] Ignore locked minidumps
Makes Android and CrOS ignore files ending with .lock or that have
associated .lock files when pulling minidumps. .lock files themselves
aren't useful and are prone to being deleted before we can actually pull
them. Files that have an associated .lock file are implied to not be
fully written to disk yet, and thus should not yet be pulled.

Bug: chromium:1056235
Change-Id: Ib263d606a27d69b2d4b87a16757473353947e859