pinpoint-ux: Improve readability of “Found n differences” message.
In particular:

1. List at most the top 3 culprits.

2. List those culprits *in order* (most significant difference first),
so that readers know for sure which commit pinpoint used when choosing an
assignee.

3. Replace ambiguous “after each of N” wording, which could be read as
meaning N commits were tried and they were all found to have significant
differences.  Instead the summary line simply says “at N”, and then on
the next line say how many commits were examined.  This both makes the
summary more concise, and gives a quick indication if the job gave poor
results (i.e. the number of differences found is a large fraction of the
number of commits tested, probably these results are noisy).

Bug: chromium:1069331
Change-Id: I43191962935dba5225e7f63983871c74b0c62dd4