auto-triage: Support a project arg in tracker
This change introduces a project argument to the APIs associated with
the issue tracker service. This allows users of these functions to
introduce the project as an argument.

This is a refactoring of some things, while introducing a new API
(/api/bugs/p/<project>/<bug_id>) for supporting future functionality.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1082029
Change-Id: I587c93046e3c3b32f628103b7ba358df10a6724c