Ignore .dat files when pulling CrOS dumps
Ignores .dat files in addition to .lock files when pulling minidumps
on CrOS since they're not useful and also apparently prone to being
deleted by the time we try to pull them. Also makes failures to pull
a file when pulling minidumps non-fatal, instead just logging loudly.

Additionally adds some unittests for the changes + a few that were
were missing on Android when the original locked file change went in.

Bug: chromium:1082255
Change-Id: I922aaf0e83b457dd0126bbdcf2cf9f2f82bccb5f