Update bq_rows export job for updated schema.
As part of the change to make the rows table be clustered, the schema was also
tweaked to add NOT NULL to `master` and `bot` columns, and add a new
`measurement` column.

Also update README.md for current version of gcloud tool, which now expects
subnetworks to be set via the --subnetwork option.

Bug: chromium:1082639
Change-Id: I94d13a5ec5329d2f075c400595db2e87f6a87fcf