Aggressively defer imports in sheriff_config_client.
There appears to be a nasty import ordering problem involving the
google.* packages: depending on the exact sequence of imports either
google.auth or google.protobuf fails to be found.  This means that the
test loader can fail intermittently with an ImportError, depending on
the order that it finds test modules in the filesystem.

This was reproducible (at least on my workstation) by running just
alert_groups_test:

$ ./bin/run_py_tests dashboard/alert_groups_test.py

There's a very complex set of import-time sys.path hackery involved
at many points, so it's hard to identify a precise culprit.  But this
change appears to be sufficient to make every single *_test.py module
importable in isolation again.  I've verified this with bash loop doing
(essentially) "python -c 'import $footest'" for every *_test module in
dashboard/.

Change-Id: I94bc1b551bdd8a6d6b341d366311a4b11e3f7276