auto-triage: Tidy up issue filing
This change ensures that we can handle the propagation of
monorail_project_id's from the sheriff-config service down to the actual
issue tracking service calls in the Dashboard.

A lot of this change is in tests which verifies that our expectations of
whether the monorail_project_id makes its way through the layers of
indirection. The crucial change here is in the sheriff_config_client
which does its own marshalling/conversion from what we get from the
service as JSON, into an NDB model. Because the initial deployment of
the support for monorail_project_id only changed the service-side protos
and since we aren't using the protos directly in the Dashboard code, we
end up missing the conversion of the field in the proto into the model.
This means while the rest of the code can handle the case when the
Subscription model has the monorail_project_id field, the client didn't
actually carry over that field in the conversion.

We really should consider just using the protobuf message in all the
places and remove the Subscription NDB model.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1086639
Change-Id: I83ef3575e61f348f022d21ae912eec10641de3b8