operations: Use explicit command in Swarming tasks
This change makes Pinpoint use the 'command' property when starting
Swarming tasks. We're doing this in response to the deprecation and
eventual removal of the command property in isolates. This change
includes the following:

- We introduced an extension point in the test runner quest (RunTest) to
  allow subclasses to define how the command is computed based on
  arguments provided to the Pinpoint job.

- We're integrating other kinds of tests in the 'new' handler, to allow
  folks to run Pinpoint jobs to specify browser tests, instrumentation
  tests, and other non-telemetry tests that generate histograms or
  chartjson.

- We are making backward compatible shims to ensure that new fields in
  the execution/quest classes are well represented, to avoid production
  issues.

This change also gives folks the opportunity to explicitly provide a
command which is not dependent on the target, if this becomes useful in
the future. Follow-up changes should involve allowing users to provide
an explicit target (instead of assuming we only support Telemetry
benchmarks in the UI).

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1090594
Change-Id: Ib81afa2b866fad7efc15729e195f7feb7c1f3e3f