Update patch set 14
Patch Set 14:

LGTM

I'm the right reviewer for this now. 😊

It's a long weekend here in Sydney and we have push-on-green set up. I'd like to land this when someone else is in the office even if the risks are low.

If you're keen, it would be great if we can remove these hard-coded limits to depths (if you can figure that out). I'd be happy with a more extensive refactoring so that this can be more generic.

Patch-set: 14
Reviewer: Gerrit User 1327821 <1327821@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>