Update patch set 14
Patch Set 14:

> Patch Set 14:
> 
> LGTM
> 
> I'm the right reviewer for this now. 😊
> 
> It's a long weekend here in Sydney and we have push-on-green set up. I'd like to land this when someone else is in the office even if the risks are low.
> 
> If you're keen, it would be great if we can remove these hard-coded limits to depths (if you can figure that out). I'd be happy with a more extensive refactoring so that this can be more generic.

I think the tricky part is handling the ndb.ComputedProperty in graph_data.py in a generic way - it isn't so obvious to me how to do this not being familiar with ndb. If possible, I would prefer to get this in with the hardcoded extension to solve the immediate problem my team has been hitting. But, I would be happy to look into a more generic solution if we ever hit this again.

Does that make sense?

Patch-set: 14
Reviewer: Gerrit User 1130766 <1130766@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0