Update patch set 2
Patch Set 2:

I'm not sure why updating the Chrome binaries would be causing this, but I think the failure is implying that the browser backend is getting recreated.

It's complaining that it's trying to create a duplicate artifact, which AFAIK should only be possible if the minidump isn't listed in the set that's already been symbolized https://cs.chromium.org/chromium/src/third_party/catapult/telemetry/telemetry/internal/backends/browser_backend.py?q=_SymbolizeAndLogMinidumps&sq=package:chromium&dr=CSs&l=358.

AFAIK that should only be possible if the browser backend is recreated, and thus the previous set is lost.

Patch-set: 2