Update patch set 2
Patch Set 2:

Ran a failing test locally, and while I didn't get the exact same failure, I went down the same code path and failed because the test left around an unsymbolized minidump. I confirmed that this does *not* fail without the CL applied, so it seems like the new reference binary is producing a minidump when previously it didn't. I'm not quite sure whether this is a legitimate issue (unexpected crash), or if the old one just didn't create a minidump when it would be normal (we are testing a renderer crash).

I'll dig deeper to figure out which it is + fix the issue.

There does appear to be a legitimate bug in the code that was causing the duplicate artifact issue - I can include a fix for that, as well.

Patch-set: 2