Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 2: Code-Review+1
> 
> (1 comment)
> 
> Looks good! Does this work with unreleased devices too?

The high and normal performance modes work in the same way they did before this change (i.e. it depends on the device).

The little-only mode works only for the devices listed in the perf_control.py, so no unreleased devices unfortunately. But when running local experiments, it's possible to turn big cores off manually and set --performance-mode=keep.

Patch-set: 2