Peak Gpu Memory Ablation Telemetry
This encompasses updates to telemetry to parse the new trace events that
are a part of the Peak Gpu Memory Ablation study.

The master of the Chromium change is:
https://chromium-review.googlesource.com/c/chromium/src/+/2187203

The tracing categories were added to the perf tools:
https://chromium-review.googlesource.com/c/chromium/src/+/2246931

The update to benchmarks to call this metric:
https://chromium-review.googlesource.com/c/chromium/src/+/2259487

TEST=ran rendering.desktop and loading.desktop with the above patch
adding the traces. Saw associated values
Bug: chromium:908462
Change-Id: Ib3a5864a95ffc3bf65453560f696d453c3304c3f