Update patch set 4
Patch Set 4:

(1 comment)

I'm not sure why you're changing things in this CL, so it would be good to add an explanation to the CL description (why you're changing it, not just what you're changing).

In particular, --quiet wasn't supposed to suppress *all* output. It's not clear to me what you'll still be logging (if anything) after this change.

Patch-set: 4