pinpoint: Use incremental builds for Pinpoint
Because we have exclusive builders for Pinpoint, let's leverage previous
build artifacts to help us reduce the need to rebuild as much of the
codebase as possible. When performing bisections this can typically
reduce build times by letting goma increase the cache hit rate and the
need to rebuild every time.

R=abennetts@google.com, fancl@chromium.org

Change-Id: I58a5333e7328445211267fdd5093bf796403d83b
Bug: chromium:1099176