[Telemetry] Retry CryoptohomePath on failure
Changes CrOSInterface.CryptohomePath to try the underlying command
up to 3 times. This is a workaround for some occasional SSH-related
flakes we're seeing, mostly on the VMs. The failures appear to be
transient, as following tests succeed, so hopefully these retries will
prevent the bots from failing while the root cause is investigated
further.

Bug: chromium:1091553
Change-Id: Idaec457ed43848835690ba3be04d52f741c2db00