Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> (1 comment)
> 
> From the stack trace in the bug - it looks like the failure is on shutdown? If so, we should probably ignore this failure?

It actually looks like it's at the beginning of the test, at least in https://chromium-swarm.appspot.com/task?id=4c96c2a793bea310.

It's happening during StopBrowser, but that's called by RestartBrowserIfNecessaryWithArgs, which we do at the beginning of a test if the browser args for a test are different than the ones the browser is currently launched with.

Patch-set: 1
Label: Commit-Queue=0