sched: Support per-configuration job concurrency
This introduces an option for bot configurations to support Pinpoint job
scheduling parameters. We're setting this per-configuration to allow
Pinpoint administrators to associate costs to each kind of job and allow
cost-based packing of jobs depending on the dimensions of a pool.

The initial implementation uses budgets and costs for different
comparison modes. We're changing the scheduler to be aware of using
budgets and allowing multiple jobs to be scheduled concurrently to
consume the budget. We're designing for over-subscription which means we
can exhaust the budget with our priority-based scheduler.

The costs per job are assigned at creation time, because adding
retroactive cost-based scheduling will require we make operations on the
`bot_config` module to be transaction-aware and thus unnecessarily
increase the costs throughout its uses in Pinpoint. We'll do this for
now as a work-around until we can move the bot configurations to
git-hosted files watched by luci-config instead of being UI managed
configurations served via Datastore.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:981429
Change-Id: Ida9cc2c365872f18cb8511a32e430c2c640bd1a7