[rendering] Report only the average for thread-times metrics.
The individual data points for the per-frame cpu-time/tasks metrics are
not particularly useful, and makes it difficult to use pinpoint when
doing bisects (more details in crbug.com/1073148#c9). So remove the
detailed data points, and report only the average per-frame metrics for
cpu-time/tasks.

This change also reduces the output results significantly:
 . On windows: from ~4.6MB to ~1.5MB
 . On android: from ~2.2MB to ~0.9MB

BUG=chromium:1073148

Change-Id: I50aa4c91b068f528e17bcbb2490be9d98a09802a