[Telemetry] Fix CrOS remote screenshots
Fixes remote CrOS screenshots under certain conditions, namely when
running on swarming. Previously, we were using the provided filepath
regardless of whether we were running in local or remote mode. This
caused issues with the screenshot failing due to the file not existing
or not being able to be created.

This is fixed by differentiating between the host and device paths, and
using a temporary, known-good location for the device path when running
in remote mode.

Bug: chromium:1098485, chromium:1098450
Change-Id: Idaf745d25621e02031d9ac67b89d856911d04ee4