Update patch set 2
Patch Set 2:

The code change looks fine but I guess I'm questioning if we want to have it be this way. 

go run src/wpr.go doesn't run but is the more general use case to go build wpr.go and then refer to it? The catapult uses mostly build the binary and then refer to all the .pem and other files explicitly.,,

Patch-set: 2