dashboard: Be more liberal with bisections
This change reduces the constraints placed on what points/anomalies in
the UI can be bisected. In the past we only allowed bisections on
non-single-change ranges. We're removing that constraint now so that we
can allow single-point bisections for reproducing issues.

In this change we reflect reality that there are no masters and slaves
involved in the implementation of this service. We're removing loaded
terminology and taking a stand that we don't condone and do not support
using terms that have historical baggage.

Let the record reflect that we will do better going forward.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1096626
Change-Id: Ib33921d1058006a0674e965ee8ba94cd61f8f2d1