Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> The code change looks fine but I guess I'm questioning if we want to have it be this way. 
> 
> go run src/wpr.go doesn't run but is the more general use case to go build wpr.go and then refer to it? The catapult uses mostly build the binary and then refer to all the .pem and other files explicitly.,,

There is always the use-case of running the server manually for debugging pages.
That's what 100% of V8 devs do when debugging benchmark.
So having reasonable working default values is quite important for us to remove yet another configuration issue when debugging.

Of course, no expert here in what is good :) but the goal should be to just do go run src/wpr.go with as few steps as possible.

Patch-set: 2
Label: Commit-Queue=0