Fix handling of md5sum_bin errors when running Unit tests
Situations can occur where the MD5 checksum tool fails (as covered in
crbug.com/1080856). The build scripts have handling to gracefully
fallback in these cases but it is broken in the case where everything is
completely up to date (I think) because there is a code path in
_GetChangedFiles which returns a 2 element tuple rather than the
expected 3 element tuple (the missing files element is omitted).

Fixed by adding an empty set into the correct location in the tuple.

Bug: chromium:1101750
Change-Id: Ic0584ccff43e8aa9f4d5a6657d8b5e42ae004f44