[Telemetry] Swap CrOS shutdown order
Swaps the order in which the CrOSInterface has its connection closed
and we await cryptohome to not be mounted during CrOS browser shutdown.

Previously, we were closing the connection and then immediately trying
to check that cryptohome was not mounted (which involves SSH when
running remotely), which may be the cause of some SSH issues we have
during browser shutdown.

Bug: chromium:1091553
Change-Id: I29cb2148b2668ec479eabc19f2ade2f15157a632