Rename whitelist/blacklist to allowlist/denylist in Perf Dashboard
To use more descriptive accurate and stop using the metaphorical
meaning of "black" and "white". See go/allowlist. So for example
the IP whitelist would now be called the IP allowlist.

This changes variable names etc. but doesn't change key string
literals or handler paths, which could be changed separately.

Change-Id: I1f1cb1403b1e0199e6eb313673349a6841fd5b16