Allow running perf tests with trichrome and standalone WebView
Right now WebView perf tests can only be run with standalone WebView for
Android L-M and monochrome for N+. This patch allows forcing other
versions of WebView.

Trichrome is only used on Q+, which we don't have any perf bots running
(that I know of). I added a trichrome option here too because it is
still nice to have this when running locally.

There was a small change to the install order of APKs in
android_browser_finder.py since trichrome WebView requires the
trichrome library to be installed first.

Bug: chromium: 1105096
Change-Id: I545d7f67037efa3a02893d1540694d49eaab888d