Disable WebView fallback logic when using WebView APK
On Android N-P, the WebView implementation is Monochrome by default.
There is special logic to fallback to the preinstalled WebView, but this
logic breaks sideloaded WebView APKs. This should be disabled anytime a
standalone WebView APK is used instead of Monochrome. Calling
SetWebViewFallbackLogic() is a no-op on Android versions other than N-P.

Bug: chromium:1105096
No-Try: True
Change-Id: I9157c9f5e70844290c9f753225bd8322b5fa57a0