[Telemetry] Log screenshot output on failure
Logs the screenshot capture command's stdout and stderr on CrOS if we
fail to pull the screenshot off the device. We're running into issues
with trying to pull a non-existent screenshot file in some cases, so
this will hopefully give us insight into why the screenshot is failing
to save in the first place.

Bug: chromium:1106927
Change-Id: I456d872d6da94a7a8748164eb0eb34a48ad80d55