Reduce flakiness around side loading system perfetto on pixel 2s.
When running processes will be created but never run (killed due to
process limits or pressures) to detect this we retry as we would if this
returned a visible error to adb RunShellCommand.

This removes the chance that we flake because normally this is rare for
an individual command but with 3 we were seeing the pixel 2 flake out
often.

Change-Id: I64de6b538bea47e22ca575c4b2fb953b8377290a