proj-perf-tooling: Add infoBlurb in issues
This change ensures that the `infoBlurb` diagnostic associated with
benchmarks can be provided in the issues filed by auto-triage. This
makes it possible to have more detailed information in the benchmark
definition for further diagnosis/exploration. This diagnostic can
include links to further documentation, in-line escalation information,
amongst other things.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1110776
Change-Id: I0db451a26ac6becf832cb41c84071719adb50625