[Telemetry] Add screenshot retry/timeout logic
Adds retry/timeout logic to Telemetry's screenshot code and enables it
on CrOS. We often try to capture a screenshot on failure when the
device's screen is temporarily off, which causes no actual screenshot
data to be captured. However, the screen turns on shortly afterwards, so
subsequent screenshot attempts should succeed.

Bug: chromium:1106927
Change-Id: I62beb1024e77e4eb6b8a350a3ccff8eeee455ff1