Clean up artifact implementation after tests
Cleans up artifact implementations after a test completes. Previously,
we were assuming that either:
1. All tests set an artifact implementation
2. Something else was handling test hermeticness, e.g. each test being
   run in a separate subprocess.

If neither of those were true, then we could get weird behavior with
tests being able to create artifacts when multiple tests are run but not
when run alone, or artifacts being attributed to the wrong test due to
using a stale instance.

Bug: chromium:1111474
Change-Id: I57d1fa8698caac35c3da125b5fbc6e436a974f65