Make Catapult "build" under Python3.
This is a patch for "building" NaCl as a part of Chromium under Python3.

Specifically, what this means is that you can use Python3 for *most*
of the build, in the sense that you can run
`gn gen --script-executable=python3 //out/Default &&
 ninja -C out/Default` and it should work just as a normal build
would work.

It does *not* yet mean that Python2 is not required; there are still
many steps of the build that require Python2, and you must still have
a Python 2 binary on your system. However, at least now all of
those places are known and annotated accordingly, so that we can
then track fixing them explicitly.

See https://crrev.com/c/2333868 for the roll-up Chromium patch,
which also has multiple other dependencies.


Commit: false
Bug: 81
Change-Id: Ic4766a7ff2915c822809c8d0252936325bb53305