[benchmarking] Add win-laptop tag for filtering win-10_laptop_low_end-perf
In order to skip running some stories,  Benchmarking uses tags defined in typ_expectations.py to update expectations.config. Though, we don't have a tag to allow us stop running a story only on win-10_laptop_low_end-perf.

This CL checks the machine type. If it is a laptop, we add a tag for it.

Bug: chromium:1114681
Bug: chromium:1115156
Change-Id: I1ff807f952ba74bd8c8de60377670c66aa51f663