[Telemetry] Always flush pagecache for browser directory on Android
Previously we assumed that on Android N+ the Webview usage in many apps
would make Chrome's code (dex and DSO) almost always resident.

We now have more evidence that with memory pressure on the majority of
devices the pagecache for Chrome files (dex/DSO) is washed out more than
we see on bots (currently most of them running Android O).

Adjusting the flushing by including the app directory for pagecache
flushing.

PERF SHERIFFS: This is expected to regress a few startup.mobile stories.

Bug: chromium:1119406
Change-Id: I927bc80ecbab8dcb8f71ef8d14b50d27bed3f63c