devil: Roll ADB to 1.0.41 (Version 30.0.3-6597393).
Let's see what explodes this time.

The most substantial change with 30.0.1 is that device-unreachable errors
go from:
"error: device blah not found"
to:
"adb: device blah not found"
And if you rename the adb binary to "adb-foo" it becomes:
"adb-foo: device blah not found"

So we can't exactly count on it always starting with "adb:". So this
just removes the expected "error: " prefix from the regex. This also
ensure backwards-compatibility if someone chooses to use an older ADB.

This was tested out on several android trybots without error in:
https://chromium-review.googlesource.com/c/chromium/src/+/2417253/2

Bug: chromium:1005504
Change-Id: I930d6fd547f06467a1301b778d8d8b088b4e3f25