[Devil] Add install sanity check
Adds a sanity check that the package is actually installed on the device
after requesting an APK installation. This will hopefully catch cases
where Chrome supposedly gets installed but then isn't available when we
actually try to use it, and if so, we can try to work around the issue,
e.g. with an install retry.

Bug: chromium:1132297
Change-Id: I9c0515bd3b106dc534999259c9786331ff5ed2e9