Add directories to sys.path w/ insert, not append.
When you are attempting to manually modify the PYTHONPATH
in a python file in order to be able to import a new package
from some other part of the source tree, it's important to
insert the directory at the front of the path, rather than
appending it to the end of the path.

If you do the latter, then you risk picking up a system-local
version of the module instead, which can result in surprising
problems down the road.

Bug: chromium:11335502
Change-Id: Ie5f394fd0303be22411555e33660f134369a1e86