chromeperf: Fix issue management
Recent changes to support different monorail projects have yielded some
issues with the way we're finding/associating issues to bisections and
even when yielding results to certain pages. This change attempts to
cover the gaps with the way we're maintaining the mapping between issues
filed against anomalies and how we're finding/merging/managing them.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1122554, chromium:1131908
Change-Id: I8624c1071a04d94127aa9979488ef11c1f48a538