Add msm gpu trace support
Add/fix drm based vblank recognition

The event name is "vblank:N" where N is the crtc index.  So the primary
display would be "vblank:0".


Add msm gpu trace support

Add support for drm/msm tracepoints, to extract GPU freq and see running
(and queued) duration for submits (jobs).


Add way to have a hidden container slice

I want to group two async slices, one for the time a job is queued, and
another for the time a job is running.  Subslices can achive this, but
in this case I don't want to see the container slice.

Change-Id: I96ee541f789222ac40cefa4227e4dfa21c130ca9