Use flexbox CSS layout in NetLog viewer.
The NetLog viewer was previously doing its own layout programmatically in JavaScript. This CL switches most of the absolute positioning to instead use automatic layout with CSS's flexbox.

This is mostly a no-op refactor, except for some minor differences:

(1) The source list now has a horizontal scrollbar. This is intentional and an improvement.
(2) The canvas in timeline view reacts more slowly to resizes, and there are some quirks with horizontal scrollbar. This was not intentional, but deferring fixes to a followup.


Bug: chromium:1026294
NOPRESUBMIT=true

Change-Id: I8560574cbfe0633dce0f708c612f1f43097b8438