pinpoint: Treat missing values differently
This change makes Pinpoint handle the cases where transitions from "no
values" to "some values" as separate from the cases where we find
significant differences between commits. There are three main cases:

- There are multiple differences, some no-value -> values and some
  significant differences.

- There's one significant difference and some no-value -> values cases.

- There are only no-value -> values cases.

When we handle the final case (only no-value->values cases) we treat
them as 'Pinpoint-No-Repro' cases and automatically mark issues WontFix,
assigning to a human and adding the 'Pinpoint-Multiple-MissingValues'
label.

In the other cases, we add the 'Pinpoint-Multiple-MissingValues' in
addition to messaging in the update pointing to the missing missing
values cases. When there's only one significant culprit, we will mark
those with the 'Pinpoint-Culprit-Found' label along with the
'Pinpoint-Multiple-MissingValues' label. This allows us to find the
cases where we can differentiate the kinds of multiple-culprit cases
with some additional fidelity.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1119962
Change-Id: I04bf8c6e4c7650e536d99e1141406184fa551f30