[typ] Support ignoring tags
Adds support for specifying a list of tags to ignore for the purposes of
typ expectation tag validation. This is so that tests can ignore tags
that they will never use instead of including them at the top of the
expectations file, e.g. if two tags are functionally equivalent or if
some generated tag will never be used.

Bug: chromium:1111876, chromium:1011107
Change-Id: Ied8d1f214f79eccc919eb0229a66ad4bbfda6175