pinpoint: Use correct attribute in serialisation
This change ensures that we're looking at the specific attribute
`self.started_time` instead of the secondary signal `self.started` on
whether we have a start time on a job. This is a work-around to support
older Job instances where `started_time` was not defined despite being
`started`.

R=abennetts@google.com, fancl@chromium.org

Fixed: chromium:1139767
Change-Id: Ic21ffcad41a4cde660000c86c8337da8bbc5bf60