Reduce the data serialized in _FormatAndPostBugCommentOnComplete tasks.
This should fix “The value of property "data" is longer than 1048487
bytes.” errors during _FormatAndPostBugCommentOnComplete.  The deferred
task was including the complete dicts from Change.AsDict /
GerritPatch.AsDict for every difference.  Instead just include the
unique identifiers, and let the async task refetch the required info
(which is probably cached anyway).

This required a small change to the tests to account for the different
sequencing of AsDict calls, but this led to more concise test code
anyway once I found the trick to patching unbound methods
(https://docs.python.org/3/library/unittest.mock-examples.html#mocking-unbound-methods).

This also corrects a misleading glitch in AddDifference's signature and
docstring; its `commit` param was sometimes actually a GerritPatch not a
Commit.  Now it is consistently a Change (and renamed accordingly).

Bug: chromium:1140309
Change-Id: I5adefeff0a126e1aa3ff5d562032757590432774