[devil] Only raise exception if forwarder not killed
Early returns if the attempt to kill the host forwarder returns -9 but
`ps aux` does not find any host_forwarder processes remaining since the
important thing is that the forwarder got cleaned up, not necessarily
how.

Also removes the dmesg logging that was added for debugging since it did
not contain any useful information.

Bug: chromium:1038789
Change-Id: Ic56a5c87bda9b59ea86704d034134c771bd4da2f