Reduce warning output in tests
When running minidump tests on Windows there are some warnings that are
printed every or almost every time. It is not obvious to somebody
looking at the test output that these warnings are known and expected so
they can lead to wasted time. This change suppresses three of them:

- 'Failed to import devil.android.forwarder' - this cannot import on
Windows

- 'Unable to stop ts_proxy_server gracefully' - signal.SIGINT is not
supported on Windows

- UnicodeDecodeError: 'ascii' codec can't decode byte 0x89 in position 80: ordinal not in range(128)
This error appears to be a cross-platform bug that happens when we try
to long .dmp and .png files as text. This issue is further confused by
the fact that the .dmp file extension is used both for crash dumps and
for the textual symbolization of those crash dumps. The fix is to not do
the logging for .dmp and .png files.

Presumably these warnings affect many other tests as well. These changes
should be safe for other tests, but may be incomplete.

Bug: chromium:1084931
Change-Id: I396a632ec98cd859a5841dac97bb6c510a84a1b0